- Пользователь может создавать вопрос   - Пользователь может просматривать список вопросов   - Пользователь, находясь на странице вопроса, может написать ответ на вопрос (т.е. форма нового ответа должна быть прямо на странице вопроса, без перехода на другую страницу)   - Пользователь может просматривать вопрос и ответы к нему.
  - Пользователь может войти в систему
  - Пользователь может выйти из системы
  - Пользователь может зарегистрироваться в системе

  - Только аутентифицированный пользователь может создавать вопросы и ответы
  - Автор может удалить свой вопрос или ответ, но не может удалить чужой вопрос/ответ

+++ app/controllers/questions_controller.rb
@@ -1,10 +1,15 @@
 class QuestionsController < ApplicationController
+  before_action :authenticate_user!, except: [:index, :show]
+
   expose :questions, ->{ Question.all }
   expose :question
+  expose :answer, -> { question.answers.new }
 
   def create
+    question.user = current_user
+
     if question.save
-      redirect_to question
+      redirect_to question, notice: 'Your question successfully created.'
     else
       render :new
     end
@@ -19,7 +24,7 @@ class QuestionsController < ApplicationController
   end
 
   def destroy
-    question.destroy
+    question.destroy if question.user == current_user
     redirect_to questions_path
   end
